# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FMS Log Nexus API
  description: |
    Centralized logging and monitoring platform API for FMS Log Nexus.
    
    ## Authentication
    
    The API supports two authentication methods:
    - **JWT Bearer Token**: For dashboard users and interactive sessions
    - **API Key**: For service accounts and job integrations
    
    ## Rate Limiting
    
    - Log ingestion: 10,000 requests/minute per API key
    - Dashboard queries: 1,000 requests/minute per user
    
  version: 1.0.0
  contact:
    name: FMS Log Nexus Support
    email: support@company.com
  license:
    name: Proprietary
servers:
  - url: https://lognexus.company.com/api
    description: Production server
  - url: https://localhost:5001/api
    description: Development server
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /logs:
    post:
      tags: [Logs]
      summary: Ingest single log entry
      description: Submit a single log entry for processing
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogEntryRequest'
      responses:
        '202':
          description: Log entry accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogEntryAcceptedResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Logs]
      summary: Search logs
      description: Search and retrieve log entries with filtering and pagination
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - name: jobId
          in: query
          schema:
            type: string
          description: Filter by job ID
        - name: jobExecutionId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by execution ID
        - name: serverName
          in: query
          schema:
            type: string
          description: Filter by server name
        - name: minLevel
          in: query
          schema:
            $ref: '#/components/schemas/LogLevel'
          description: Minimum log level
        - name: search
          in: query
          schema:
            type: string
          description: Full-text search in message
        - name: correlationId
          in: query
          schema:
            type: string
          description: Filter by correlation ID
        - name: hasException
          in: query
          schema:
            type: boolean
          description: Filter entries with/without exceptions
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
      responses:
        '200':
          description: Log entries retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogEntriesResponse'
  /logs/batch:
    post:
      tags: [Logs]
      summary: Ingest batch of log entries
      description: Submit multiple log entries (up to 1000) in a single request
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogBatchRequest'
      responses:
        '202':
          description: Batch accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogBatchResponse'
  /logs/{id}:
    get:
      tags: [Logs]
      summary: Get log entry details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Log entry details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogEntryDetail'
        '404':
          description: Log entry not found
  /logs/statistics:
    get:
      tags: [Logs]
      summary: Get log statistics
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - name: jobId
          in: query
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [hour, day, job, server]
            default: hour
      responses:
        '200':
          description: Log statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogStatistics'

  # ============================================================================
  # Jobs Endpoints
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for user authentication

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service authentication
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
      description: Items per page
    SortBy:
      name: sortBy
      in: query
      schema:
        type: string
        default: timestamp
      description: Field to sort by
    SortOrder:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort order
    StartDate:
      name: startDate
      in: query
      schema:
        type: string
        format: date-time
      description: Start of time range (ISO 8601)
    EndDate:
      name: endDate
      in: query
      schema:
        type: string
        format: date-time
      description: End of time range (ISO 8601)
  schemas:
...
    LogLevel:
      type: string
      enum: [Trace, Debug, Information, Warning, Error, Critical]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
        totalCount:
          type: integer
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            traceId:
              type: string

    # ========== Auth ==========
...
    LogEntryRequest:
      type: object
      required: [level, message, serverName]
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          $ref: '#/components/schemas/LogLevel'
        message:
          type: string
          maxLength: 4000
        messageTemplate:
          type: string
        jobId:
          type: string
        jobExecutionId:
          type: string
          format: uuid
        serverName:
          type: string
        category:
          type: string
        correlationId:
          type: string
        exception:
          type: object
          properties:
            type:
              type: string
            message:
              type: string
            stackTrace:
              type: string
        properties:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
    LogEntryAcceptedResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        message:
          type: string
    LogBatchRequest:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LogEntryRequest'
          maxItems: 1000
    LogBatchResponse:
      type: object
      properties:
        accepted:
          type: integer
        rejected:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              field:
                type: string
              message:
                type: string
    LogEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        level:
          $ref: '#/components/schemas/LogLevel'
        levelName:
          type: string
        message:
          type: string
        jobId:
          type: string
        jobExecutionId:
          type: string
          format: uuid
        serverName:
          type: string
        category:
          type: string
        correlationId:
          type: string
        exceptionType:
          type: string
        exceptionMessage:
          type: string
        hasException:
          type: boolean
        properties:
          type: object
        tags:
          type: array
          items:
            type: string
    LogEntryDetail:
      allOf:
        - $ref: '#/components/schemas/LogEntry'
        - type: object
          properties:
            messageTemplate:
              type: string
            jobDisplayName:
              type: string
            serverDisplayName:
              type: string
            sourceContext:
              type: string
            traceId:
              type: string
            spanId:
              type: string
            exception:
              type: object
              properties:
                type:
                  type: string
                message:
                  type: string
                stackTrace:
                  type: string
                source:
                  type: string
            relatedLogs:
              type: array
              items:
                $ref: '#/components/schemas/LogEntry'
    LogEntriesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'
        pagination:
          $ref: '#/components/schemas/Pagination'
    LogStatistics:
      type: object
      properties:
        summary:
          type: object
          properties:
            totalLogs:
              type: integer
            traceCount:
              type: integer
            debugCount:
              type: integer
            infoCount:
              type: integer
            warningCount:
              type: integer
            errorCount:
              type: integer
            criticalCount:
              type: integer
            exceptionCount:
              type: integer
        trend:
          type: array
          items:
            type: object
            properties:
              bucket:
                type: string
                format: date-time
              total:
                type: integer
              info:
                type: integer
              warning:
                type: integer
              error:
                type: integer
        topExceptions:
          type: array
          items:
            type: object
            properties:
              exceptionType:
                type: string
              count:
                type: integer
              affectedJobs:
                type: integer

    # ========== Jobs ==========
...